Introduced __ATTRIBBUTE__.
When used as a default initializer this will resolve to either [], or (if it is within an UDA declaration)
an string literal of all the declarations the UDA is attached to. A new default init class is introduced (AttributeInitExp)
To facicilitate this a new field has been added to Scope to keep track of which user defined attribute it is attached to.
To find if we are in a UDA, resolveLoc() walks the given scope upwards until it's parent is either not valid is
from a uda declaration.